Making the content and undo views INVISIBLE instead of GONE prevents some strange layout issues.
If the 'content' or the 'undo' view is made GONE, the other view's layout changes because views that are GONE
are never part of any measuring. E.g. swiping away the 'content' view can make the entire ContextualIUndoView less high
if the 'undo' view has a different measurement that yields a lower height than the 'content's height.

 When these two views are made 'INVISIBLE' instead, the measurement of the ContextualUndoView never changes before, during
 or after a swipe and undo action.

 Also, the animation seems to be smoother when using INVISIBLE instead of GONE.