The 'Current Removed View' was never permanently removed (Android 4.3)
The 'Current Removed View' was assigned in the Adapter's getView() method.
However, the parent of this view was never set. Instead, items from the listview's 'activeViews' list were used
instead and they never went through the 'getView()' for measuring again.

Changed this by adding an onHierarchyChangeListener and picking up in the 'Current Removed View' in the 'onChildViewAdded' method.